refactor: 입력 다시 받기 기능 추가
